+++ src/Ethernet_Generic.h
@@ -32,7 +32,7 @@
   OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
   SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-  Version: 2.8.0
+  Version: 2.8.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -44,6 +44,7 @@
   2.7.0   K Hoang      14/11/2022 Fix severe limitation to permit sending larger data than 2/4/8/16K buffer
   2.7.1   K Hoang      15/11/2022 Auto-detect W5x00 and settings to set MAX_SIZE to send
   2.8.0   K Hoang      27/12/2022 Add support to W6100 using IPv4
+  2.8.1   K Hoang      06/01/2023 Fix W6100 minor bug
  *****************************************************************************************************************************/
 
 #pragma once
@@ -54,13 +55,13 @@
 
 ////////////////////////////////////////
 
-#define ETHERNET_GENERIC_VERSION            "Ethernet_Generic v2.8.0"
+#define ETHERNET_GENERIC_VERSION            "Ethernet_Generic v2.8.1"
 
 #define ETHERNET_GENERIC_VERSION_MAJOR      2
 #define ETHERNET_GENERIC_VERSION_MINOR      8
-#define ETHERNET_GENERIC_VERSION_PATCH      0
+#define ETHERNET_GENERIC_VERSION_PATCH      1
 
-#define ETHERNET_GENERIC_VERSION_INT        2008000
+#define ETHERNET_GENERIC_VERSION_INT        2008001
 
 ////////////////////////////////////////
 
@@ -76,12 +77,6 @@
   #if(_ETG_LOGLEVEL_> 3)
     #warning ETHERNET_GENERIC_USING_SPI in Ethernet_Generic.h
   #endif
-#endif
-
-////////////////////////////////////////
-
-#if !defined(USING_W6100)
-  #define USING_W6100       true
 #endif
 
 ////////////////////////////////////////