+++ src/utility/w5100_Impl.h
@@ -13,7 +13,7 @@
 
   Built by Khoi Hoang https://github.com/khoih-prog/EthernetWebServer
 
-  Version: 2.8.0
+  Version: 2.8.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -25,6 +25,7 @@
   2.7.0   K Hoang      14/11/2022 Fix severe limitation to permit sending larger data than 2/4/8/16K buffer
   2.7.1   K Hoang      15/11/2022 Auto-detect W5x00 and settings to set MAX_SIZE to send
   2.8.0   K Hoang      27/12/2022 Add support to W6100 using IPv4
+  2.8.1   K Hoang      06/01/2023 Fix W6100 minor bug
  *****************************************************************************************************************************/
 
 #pragma once
@@ -651,6 +652,8 @@ uint8_t W5100Class::softReset()
 uint8_t W5100Class::isW6100()
 {
   chip = w6100;
+  altChip = noChip;
+  
   CH_BASE_MSB = 0x80;
 
   ETG_LOGDEBUG("isW6100: detect W6100 chip");
@@ -1426,7 +1429,6 @@ const char* W5100Class::linkReport()
   }
 
 #if USING_W6100
-  // KH TODO verify for W6100
   else if (chip == w6100)
   {
     if (bitRead(readPHYCFGR_W6100(), 0) == 1)
@@ -1470,12 +1472,13 @@ const char* W5100Class::speedReport()
   }
 
 #if USING_W6100
-  // KH TODO verify for W6100
+	// Check https://docs.wiznet.io/img/products/w6100/w6100_ds_v105e.pdf, page 37
+	// Inversed from W5x00
   else if (chip == w6100)
   {
     if (bitRead(readPHYCFGR_W6100(), 0) == 1)
     {
-      if (bitRead(readPHYCFGR_W6100(), 1) == 1)
+      if (bitRead(readPHYCFGR_W6100(), 1) == 0)
         return "100 MB";
       else
         return "10 MB";
@@ -1520,15 +1523,16 @@ const char* W5100Class::duplexReport()
   }
 
 #if USING_W6100
-  // KH TODO verify for W6100
+	// Check https://docs.wiznet.io/img/products/w6100/w6100_ds_v105e.pdf, page 37
+	// Inversed from W5x00
   else if (chip == w6100)
   {
     if (bitRead(readPHYCFGR_W6100(), 0) == 1)
     {
-      if (bitRead(readPHYCFGR_W6100(), 2) == 1)
+      if (bitRead(readPHYCFGR_W6100(), 2) == 0)
         return "FULL DUPLEX";
 
-      if (bitRead(getPHYCFGR(), 2) == 0)
+      if (bitRead(getPHYCFGR(), 2) == 1)
         return "HALF DUPLEX";
     }